refactor: replace 'common' module with 'tokens'
this better reflects what we are actually extracting and allows us to
move some other code out of the 'client' module.
